requirements.txt
flask
requests
supabase
python-dotenv
requests_toolbelt